pyproject.toml
[project]
name = "uk-osm-processor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.13.5",
    "geoalchemy2>=0.18.0",
    "geopandas>=1.1.1",
    "osmium>=4.1.1",
    "pandas>=2.3.2",
    "psycopg2-binary>=2.9.10",
    "pytest>=8.4.2",
    "pyyaml>=6.0.2",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.43",
]
